requirements.txt
absl-py==2.1.0
apache-beam==2.58.1
array_record==0.5.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
comm==0.2.2
contourpy==1.2.1
crcmod==1.7
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
dill==0.3.1.1
dm-reverb==0.14.0
dm-tree==0.1.8
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
etils==1.7.0
exceptiongroup==1.2.2
executing==2.0.1
fastavro==1.9.5
fasteners==0.19
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.4.0
gin-config==0.5.0
google-api-core==2.19.1
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.5
gym==0.23.0
gym-notices==0.0.8
h5py==3.11.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.7
immutabledict==4.2.0
importlib_resources==6.4.3
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Js2Py==0.74
jsonpickle==3.2.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
objsize==0.7.0
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
portpicker==1.6.0
promise==2.3
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydot==1.4.2
pygame==2.1.3
Pygments==2.18.0
pyjsparser==2.7.1
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.1
redis==5.0.8
referencing==0.35.1
regex==2024.7.24
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rlds==0.1.8
rpds-py==0.20.0
rsa==4.9
simple_parsing==0.1.5
six==1.16.0
stack-data==0.6.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.1
tensorflow-datasets==4.9.6
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.15.0
tensorflow-probability==0.22.0
termcolor==2.4.0
tf-agents==0.19.0
tfp-nightly==0.25.0.dev20240820
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.14.1
zipp==3.20.0
zstandard==0.23.0
